refactor(data/remotegetuser): move files to your folders
