+++ src/data/errors/index.ts
@@ -1,2 +1,3 @@
 export { default as ValidationError } from './validationError';
 export { default as UnexpectedError } from './unexpectedError';
+export { default as UserNotFoundError } from './userNotFoundError';
