+++ src/data/http/httpGetClient.ts
@@ -1,0 +1,6 @@
+import HttpRequest from './httpRequest';
+import HttpResponse from './httpResponse';
+
+export default interface HttpGetClient<R = any> {
+  get(data: HttpRequest): Promise<HttpResponse<R>>;
+}
