+++ src/data/http/index.ts
@@ -1,4 +1,5 @@
 export { default as HttpPostClient } from './httpPostClient';
+export { default as HttpGetClient } from './httpGetClient';
 export { default as HttpRequest } from './httpRequest';
 export { default as HttpResponse } from './httpResponse';
 export { default as HttpStatusCode } from './httpStatusCode';
