+++ src/data/useCases/index.ts
@@ -1,1 +1,2 @@
 export { default as RemoteAddAccount } from './remoteAddAccount';
+export { default as RemoteGetUser } from './remoteGetUser';
