+++ src/domain/useCases/index.ts
@@ -1,1 +1,2 @@
 export { default as AddAccount } from './addAccount';
+export { default as GetUser } from './getUser';
