+++ tests/data/http/httpClientSpy.ts
@@ -1,17 +1,27 @@
 import {
+  HttpGetClient,
   HttpPostClient,
   HttpRequest,
   HttpResponse,
   HttpStatusCode,
 } from '~/data/http';
+import { UserModel } from '~/domain/models';
 
-export class HttpClientSpy implements HttpPostClient {
+export class HttpClientSpy implements HttpPostClient, HttpGetClient {
+  userId = '';
   url = '';
   body: any;
   response: HttpResponse<any> = { statusCode: HttpStatusCode.ok, body: {} };
   async post(data: HttpRequest): Promise<HttpResponse<any>> {
     this.url = data.url;
     this.body = data.body;
+    return this.response;
+  }
+
+  async get(data: HttpRequest): Promise<HttpResponse<UserModel>> {
+    const urlArray = data.url.split('/');
+    this.userId = urlArray[urlArray.length - 1];
+    this.url = data.url;
     return this.response;
   }
 
@@ -23,7 +33,14 @@ export class HttpClientSpy implements HttpPostClient {
     this.response.statusCode = HttpStatusCode.forbidden;
   }
 
-  completeWithSuccess() {
-    this.response = { statusCode: HttpStatusCode.created, body: {} };
+  completeWithSuccess(body?: any) {
+    this.response = { statusCode: HttpStatusCode.created, body };
+  }
+
+  completeWithUserNotFound() {
+    this.response = {
+      statusCode: HttpStatusCode.notFound,
+      body: null,
+    };
   }
 }
